Hotfix(integration): Fix argosy for MHRise (#494)
* Fix(Core): Add support for MHRise v16.0.2.0

* Hotfix(Core): Fixed MHRise Argosy

+++ HunterPie.Integrations/Datasources/MonsterHunterRise/Definitions/MHRSubmarineStructure.cs
@@ -18,5 +18,5 @@ public struct MHRSubmarineStructure
     public int DaysLeft;
 
     [FieldOffset(0x40)]
-    public int ItemArrayPtr;
+    public long ItemArrayPtr;
 }
